+++ lib/src/priceLine/usePriceLine.test.tsx
@@ -2,6 +2,7 @@ import { renderHook } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { useSafeContext } from "@/_shared/useSafeContext";
 import { usePriceLine } from "./usePriceLine";
+import type { PriceLineProps } from "./types";
 
 vi.mock("@/_shared/useSafeContext");
 
@@ -57,5 +58,73 @@ describe("usePriceLine", () => {
     unmount();
 
     expect(mockRemovePriceLine).toHaveBeenCalled();
+  });
+
+  it("should apply options to the price line", () => {
+    vi.mocked(useSafeContext).mockReturnValue({
+      seriesApiRef: mockSeries,
+      isReady: true,
+    });
+
+    const { rerender } = renderHook(
+      props =>
+        usePriceLine({
+          price: props.price,
+          options: props.options,
+        }),
+      {
+        initialProps: {
+          price: 100,
+          options: {
+            color: "blue",
+          },
+        } as PriceLineProps,
+      }
+    );
+
+    const newOptions = {
+      color: "red",
+    };
+
+    rerender({
+      price: 100,
+      options: newOptions,
+    });
+
+    expect(mockCreatePriceLine().applyOptions).toHaveBeenCalledWith(newOptions);
+  });
+
+  it("should not create a price line if not ready", () => {
+    vi.mocked(useSafeContext).mockReturnValue({
+      seriesApiRef: mockSeries,
+      isReady: false,
+    });
+
+    const { result } = renderHook(() =>
+      usePriceLine({
+        price: 100,
+      })
+    );
+
+    const api = result.current.current.api();
+    expect(api).toBeNull();
+    expect(mockCreatePriceLine).not.toHaveBeenCalled();
+  });
+
+  it("should not create a price line if no seriesApiRef", () => {
+    vi.mocked(useSafeContext).mockReturnValue({
+      seriesApiRef: null,
+      isReady: true,
+    });
+
+    const { result } = renderHook(() =>
+      usePriceLine({
+        price: 100,
+      })
+    );
+
+    const api = result.current.current.api();
+    expect(api).toBeNull();
+    expect(mockCreatePriceLine).not.toHaveBeenCalled();
   });
 });