+++ lib/src/scales/usePriceScale.test.tsx
@@ -1,0 +1,91 @@
+import { renderHook } from "@testing-library/react";
+import { beforeEach, describe, expect, it, vi } from "vitest";
+import { useSafeContext } from "@/_shared/useSafeContext";
+import { usePriceScale } from "./usePriceScale";
+
+vi.mock("@/_shared/useSafeContext");
+
+const mockApplyOptions = vi.fn();
+
+const mockChart = {
+  api: () => ({
+    priceScale: vi.fn().mockReturnValue({
+      applyOptions: mockApplyOptions,
+      options: vi.fn(),
+      width: vi.fn(),
+    }),
+  }),
+};
+
+describe("usePriceScale", () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  it("initializes priceScale", () => {
+    vi.mocked(useSafeContext).mockReturnValue({
+      isReady: true,
+      chartApiRef: mockChart,
+    });
+
+    const { result } = renderHook(() =>
+      usePriceScale({
+        id: "right",
+      })
+    );
+
+    const api = result.current.current.api();
+    expect(api).toBeDefined();
+  });
+
+  it("applies options to priceScale", () => {
+    vi.mocked(useSafeContext).mockReturnValue({
+      isReady: true,
+      chartApiRef: mockChart,
+    });
+
+    const { rerender } = renderHook(
+      props =>
+        usePriceScale({
+          id: "right",
+          options: props.options,
+        }),
+      {
+        initialProps: {
+          options: {
+            autoScale: true,
+          },
+        },
+      }
+    );
+
+    expect(mockApplyOptions).toHaveBeenCalledWith({
+      autoScale: true,
+    });
+
+    rerender({
+      options: {
+        autoScale: false,
+      },
+    });
+
+    expect(mockApplyOptions).toHaveBeenCalledWith({
+      autoScale: false,
+    });
+  });
+
+  it("does not initialize priceScale if not ready", () => {
+    vi.mocked(useSafeContext).mockReturnValue({
+      isReady: false,
+      chartApiRef: mockChart,
+    });
+
+    const { result } = renderHook(() =>
+      usePriceScale({
+        id: "right",
+      })
+    );
+
+    expect(result.current.current.api()).toBeNull();
+  });
+});