+++ lib/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Feat
 - add chart `onDblClick` prop to handle double-click event subscription
+### Fix
+- fix `PriceLine` price property change does not apply to price line on the chart
 
 ## [0.3.4] - 2025-05-09
 ### Fix